Implementa gerenciamento de teclas de atalho globais e melhorias em classes de áudio
- Adiciona a classe AppHotKeyManager para gerenciar teclas de atalho globais no aplicativo.
- Implementa a classe HotKeyService para registrar e processar teclas de atalho.
- Atualiza a classe GerenciadorEfeitos para implementar IDisposable e liberar recursos.
- Atualiza a classe EqualizadorAudio para implementar IDisposable e liberar recursos.
- Atualiza a classe AudioVisualizer para implementar IDisposable e liberar recursos.
- Atualiza a classe Reprodutor para implementar IDisposable e liberar recursos.
- Adiciona suporte a teclas de atalho globais na página de configurações.
- Adiciona interface de usuário para habilitar/desabilitar teclas de atalho globais na SettingsPage.
- Melhora a manipulação de volume e reprodução nas classes relacionadas ao áudio.

+++ Fulcrum/View/HomePage.xaml.cs
@@ -301,8 +301,8 @@ public sealed partial class HomePage : Page
         
         try
         {
-            // Atualiza o volume do reprodutor correspondente
-            AudioManager.Instance.AlterarVolume(soundId, e.NewValue);
+            // Atualiza o volume do reprodutor correspondente - convertendo explicitamente para float
+            AudioManager.Instance.AlterarVolume(soundId, (float)e.NewValue);
         }
         catch (Exception ex)
         {
@@ -564,7 +564,7 @@ public sealed partial class HomePage : Page
             // Ajusta o volume para todos os reprodutores
             foreach (var reprodutor in AudioManager.Instance.GetListReprodutores())
             {
-                AudioManager.Instance.AlterarVolume(reprodutor.Key, slider.Value);
+                AudioManager.Instance.AlterarVolume(reprodutor.Key, (float)slider.Value);
             }
         }
     }
